Server/pyproject.toml
[project]
name = "server"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.5",
    "firebase-admin>=6.6.0",
    "google-cloud-core>=2.4.1",
    "google-cloud>=0.34.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.32.1",
    "google-cloud-firestore>=2.19.0",
    "pytz>=2024.2",
    "toml>=0.10.2",
    "pydantic[email]>=2.10.2",
    "qrcode>=8.0",
    "pillow>=11.0.0",
    "fabric>=3.2.2",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.19",
    "requests>=2.32.3",
]
